+++ src/Commands/Solr/DisableCommand.php
@@ -27,9 +27,17 @@ class DisableCommand extends TerminusCommand implements SiteAwareInterface
      * @param string $site_id Site name
      *
      * @usage <site> Disables Solr add-on for <site>.
+     *
+     * @deprecated Use 'search:disable --flavor=solr' instead. This command will be removed in a future version.
      */
     public function disable($site_id)
     {
+        $this->log()->warning(
+            'The "solr:disable" command is deprecated. ' .
+            'Please use "search:disable --flavor=solr" instead. ' .
+            'This command will be removed in a future version of Terminus.'
+        );
+
         $site = $this->getSiteById($site_id);
         $workflow = $site->getSolr()->disable();
         $this->processWorkflow($workflow);