+++ src/Plugins/PluginDiscovery.php
@@ -27,6 +27,8 @@ class PluginDiscovery implements
      */
     public const BLACKLIST = [
         'pantheon-systems/terminus-aliases-plugin',
+        'pantheon-systems/terminus-node-logs-plugin',
+        'pantheon-systems/terminus-repository-plugin',
         'pantheon-systems/terminus-secrets-manager-plugin',
     ];